Refactor : Work 중복 type 제거 및 Skills type과 Education type 모두 about.ts 파일로 이동
