+++ src/components/work/Work.tsx
@@ -1,13 +1,15 @@
 import { getWorkList } from '@/lib/api/workApi'
 import WorkList from './WorkList'
-import { WorkItemData } from '@/types/workList'
+import { WorkItemPropData } from '@/types/works'
 
 export default async function Work() {
   const workList = await getWorkList()
 
   return (
     <>
-      {!!workList?.length && <WorkList dataList={workList as WorkItemData[]} />}
+      {!!workList?.length && (
+        <WorkList dataList={workList as WorkItemPropData[]} />
+      )}
     </>
   )
 }
