+++ src/types/about.ts
@@ -1,3 +1,20 @@
+// Skills
+export type SkillItemData = {
+  properties?: {
+    Name: {
+      title?: {
+        plain_text?: string
+      }[]
+    }
+    Type: {
+      select?: {
+        name?: string
+      }
+    }
+  }
+}
+
+// Education
 export type EducationItemData = {
   properties?: {
     Name: {
