--- src/types/skills.ts
@@ -1,14 +1,0 @@
-export type SkillItemData = {
-  properties?: {
-    Name: {
-      title?: {
-        plain_text?: string
-      }[]
-    }
-    Type: {
-      select?: {
-        name?: string
-      }
-    }
-  }
-}
